services/api-gateway/src/api_gateway/security_headers.py
"""HTTP-Security-Header fuer alle Gateway-Responses (ohne HTML-CSP — reine JSON-API)."""

from __future__ import annotations

from collections.abc import Callable

from starlette.middleware.base import BaseHTTPMiddleware
from starlette.requests import Request
from starlette.responses import Response

from api_gateway.config import get_gateway_settings


class SecurityHeadersMiddleware(BaseHTTPMiddleware):
    async def dispatch(self, request: Request, call_next: Callable) -> Response:
        response = await call_next(request)
        settings = get_gateway_settings()
        response.headers.setdefault("X-Content-Type-Options", "nosniff")
        response.headers.setdefault("X-Frame-Options", "DENY")
        response.headers.setdefault(
            "Referrer-Policy",
            "strict-origin-when-cross-origin",
        )
        _perm = (
            "accelerometer=(), camera=(), geolocation=(), gyroscope=(), "
            "magnetometer=(), microphone=(), payment=(), usb=()"
        )
        response.headers.setdefault("Permissions-Policy", _perm)
        if settings.gateway_send_hsts:
            response.headers.setdefault(
                "Strict-Transport-Security",
                "max-age=31536000; includeSubDomains",
            )
        csp = str(
            getattr(settings, "gateway_content_security_policy", "") or ""
        ).strip()
        if csp:
            response.headers.setdefault("Content-Security-Policy", csp)
        return response
