services/api-gateway/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "api-gateway"
version = "0.1.0"
description = "FastAPI API-Gateway: Dashboard-BFF, Auth, Rate-Limits, Live-SSE, Live-Broker-Proxies."
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "fpdf2==2.8.7",
  "httpx",
  "PyJWT>=2.8.0",
  "stripe>=10.0.0",
  "prometheus_client>=0.20.0",
  "psycopg[binary]",
  "pydantic>=2",
  "pydantic-settings>=2",
  "python-json-logger>=2.0.7",
  "redis",
  "sse-starlette",
  "uvicorn[standard]",
]

[tool.hatch.build.targets.wheel]
packages = ["src/api_gateway"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = ["src"]
